데이터 중심 설계: 예매 함수 분리
